Use QStandardPaths for log filepaths
