+++ src/icarus_v2/resources/setup/70-dataq-di-4108.rules
